Refactor MCP internal handler and routes; implement Manifold MCP server
- Removed the RegisterRoutes method from InternalMCPHandler and directly registered routes in registerMCPEndpoints.
- Added nil check for mcpMgr in Close method of InternalMCPHandler.
- Created build script for the Manifold MCP server.
- Implemented main entry point for the Manifold MCP server, including signal handling and tool registration.
- Added various tools for file system operations, Git operations, and additional utilities in tools.go.
- Defined argument types for each tool to facilitate structured input.
