+++ config.yaml.example
@@ -78,3 +78,9 @@ mcpServers:
       - ghcr.io/github/github-mcp-server
     env:
       GITHUB_PERSONAL_ACCESS_TOKEN: "<YOUR_TOKEN>"
+  
+  # Manifold's internal MCP server
+  manifold:
+    command: ./cmd/mcp-manifold/mcp-manifold
+    args: []
+    env: {}
